+++ optuna/samplers/_search_space/__init__.py
@@ -1,10 +1,12 @@
 from optuna.samplers._search_space.group_decomposed import _GroupDecomposedSearchSpace
+from optuna.samplers._search_space.group_decomposed import _SearchSpaceGroup
 from optuna.samplers._search_space.intersection import intersection_search_space
 from optuna.samplers._search_space.intersection import IntersectionSearchSpace
 
 
 __all__ = [
     "_GroupDecomposedSearchSpace",
+    "_SearchSpaceGroup",
     "IntersectionSearchSpace",
     "intersection_search_space",
 ]
